refactor: MemberPrincipal 제거 [#46]
