+++ src/main/java/com/nhnacademy/member_server/controller/AddressController.java
@@ -3,20 +3,12 @@ package com.nhnacademy.member_server.controller;
 import com.nhnacademy.member_server.dto.request.member.AddressRequest;
 import com.nhnacademy.member_server.dto.response.member.AddressListResponse;
 import com.nhnacademy.member_server.dto.response.member.AddressResponse;
-import com.nhnacademy.member_server.entity.MemberPrincipal;
 import com.nhnacademy.member_server.service.member.AddressService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/address")
@@ -28,25 +20,22 @@ public class AddressController {
 
     //해당 멤버 기본 주소 불러오기
     @GetMapping("/default")
-    public ResponseEntity<AddressResponse> getDefaultAddress(@AuthenticationPrincipal MemberPrincipal principal) {
-        Long memberId = principal.getMemberId();
+    public ResponseEntity<AddressResponse> getDefaultAddress(@RequestHeader(name = "X-User-ID") Long memberId) {
         AddressResponse addressResponse = addressService.findDefaultAddress(memberId);
         return ResponseEntity.ok(addressResponse);
     }
 
     //해당 멤버 모든 주소 불러오기
     @GetMapping
-    public ResponseEntity<AddressListResponse> getAddressList(@AuthenticationPrincipal MemberPrincipal principal) {
-        Long memberId = principal.getMemberId();
+    public ResponseEntity<AddressListResponse> getAddressList(@RequestHeader(name = "X-User-ID") Long memberId) {
         AddressListResponse addressListResponse = addressService.findAddressList(memberId);
         return ResponseEntity.ok(addressListResponse);
     }
 
     //특정 주소 정보 불러오기 (address Id 기준)
     @GetMapping("/{address-id}")
-    public ResponseEntity<AddressResponse> getAddress(@AuthenticationPrincipal MemberPrincipal principal,
+    public ResponseEntity<AddressResponse> getAddress(@RequestHeader(name = "X-User-ID") Long memberId,
                                                        @PathVariable("address-id") Long addressId) {
-        Long memberId = principal.getMemberId();
         AddressResponse addressResponse = addressService.findAddress(memberId, addressId);
         return ResponseEntity.ok(addressResponse);
     }
@@ -54,36 +43,32 @@ public class AddressController {
 
     //주소 추가하기
     @PostMapping
-    public ResponseEntity<AddressResponse> createAddress(@AuthenticationPrincipal MemberPrincipal principal,
+    public ResponseEntity<AddressResponse> createAddress(@RequestHeader(name = "X-User-ID") Long memberId,
                                                          @Valid @RequestBody AddressRequest addressRequest) {
-        Long memberId = principal.getMemberId();
         AddressResponse addressResponse = addressService.registerAddress(memberId, addressRequest);
         return ResponseEntity.ok(addressResponse);
     }
 
     @PostMapping("/{address-id}/default")
-    public ResponseEntity<AddressResponse> setDefaultAddress(@AuthenticationPrincipal MemberPrincipal principal,
+    public ResponseEntity<AddressResponse> setDefaultAddress(@RequestHeader(name = "X-User-ID") Long memberId,
                                                              @PathVariable("address-id") Long addressId) {
-        Long memberId = principal.getMemberId();
         AddressResponse addressResponse = addressService.setDefaultAddress(memberId, addressId);
         return ResponseEntity.ok(addressResponse);
     }
 
     //특정 주소 업데이트
     @PatchMapping("/{address-id}")
-    public ResponseEntity<AddressResponse> updateAddress(@AuthenticationPrincipal MemberPrincipal principal,
+    public ResponseEntity<AddressResponse> updateAddress(@RequestHeader(name = "X-User-ID") Long memberId,
                                                          @RequestBody AddressRequest addressRequest,
                                                          @PathVariable("address-id") Long addressId) {
-        Long memberId = principal.getMemberId();
         AddressResponse addressResponse = addressService.modifyAddress(memberId, addressId, addressRequest);
         return ResponseEntity.ok(addressResponse);
     }
 
     //특정 주소 삭제
     @DeleteMapping("/{address-id}")
-    public ResponseEntity<Void> deleteAddress(@AuthenticationPrincipal MemberPrincipal principal,
+    public ResponseEntity<Void> deleteAddress(@RequestHeader(name = "X-User-ID") Long memberId,
                                               @PathVariable("address-id") Long addressId) {
-        Long memberId = principal.getMemberId();
         addressService.removeAddress(memberId, addressId);
         return ResponseEntity.ok().build();
     }
