+++ src/main/java/com/nhnacademy/member_server/controller/InternalController.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.member_server.controller;
 
-import com.nhnacademy.member_server.entity.Role;
+import com.nhnacademy.member_server.entity.member.Role;
 import com.nhnacademy.member_server.service.member.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
