+++ src/main/java/com/nhnacademy/member_server/controller/MemberController.java
@@ -3,8 +3,7 @@ package com.nhnacademy.member_server.controller;
 import com.nhnacademy.member_server.dto.request.member.MemberUpdateRequest;
 import com.nhnacademy.member_server.dto.response.member.MemberResponse;
 import com.nhnacademy.member_server.dto.response.member.SimpleMemberResponse;
-import com.nhnacademy.member_server.entity.MemberPrincipal;
-import com.nhnacademy.member_server.entity.Role;
+import com.nhnacademy.member_server.entity.member.Role;
 import com.nhnacademy.member_server.global.jwt.WebUtils;
 import com.nhnacademy.member_server.service.member.AuthService;
 import com.nhnacademy.member_server.service.member.MemberService;
@@ -13,7 +12,6 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -35,25 +33,22 @@ public class MemberController {
     private final AuthService authService;
 
     @GetMapping("/me")
-    public ResponseEntity<MemberResponse> getMember(@AuthenticationPrincipal MemberPrincipal principal) {
-        Long memberId = principal.getMemberId();
+    public ResponseEntity<MemberResponse> getMember(@RequestHeader(name = "X-User-ID") Long memberId){
         MemberResponse response = memberService.getMember(memberId);
         return ResponseEntity.ok(response);
     }
 
     @PatchMapping("/me")
-    public ResponseEntity<MemberResponse> updateMember(@AuthenticationPrincipal MemberPrincipal principal,
+    public ResponseEntity<MemberResponse> updateMember(@RequestHeader(name = "X-User-ID") Long memberId,
                                                        @Valid @RequestBody MemberUpdateRequest memberUpdateRequest) {
-        Long memberId = principal.getMemberId();
         MemberResponse memberResponse = memberService.updateMember(memberId, memberUpdateRequest);
         return ResponseEntity.ok(memberResponse);
     }
 
 
     @DeleteMapping("me/withdraw")
-    public ResponseEntity<Void> withdraw(@AuthenticationPrincipal MemberPrincipal principal,
+    public ResponseEntity<Void> withdraw(@RequestHeader(name = "X-User-ID") Long memberId,
                                          @RequestHeader(HttpHeaders.AUTHORIZATION) String bearerHeader) {
-        Long memberId = principal.getMemberId();
         authService.logout(WebUtils.getToken(bearerHeader), memberId);
         memberService.withdraw(memberId);
