--- src/main/java/com/nhnacademy/member_server/dto/response/PaycoTokenResponse.java
@@ -1,21 +1,0 @@
-package com.nhnacademy.member_server.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class PaycoTokenResponse {
-    @JsonProperty("access_token")
-    private String accessToken;
-
-    @JsonProperty("refresh_token")
-    private String refreshToken;
-
-    @JsonProperty("token_type")
-    private String tokenType;
-
-    @JsonProperty("expires_in")
-    private String expiresIn;
-}
