+++ src/main/java/com/nhnacademy/member_server/dto/response/member/MemberResponse.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.member_server.dto.response.member;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.nhnacademy.member_server.entity.Member;
+import com.nhnacademy.member_server.entity.member.Member;
 import jakarta.validation.constraints.Email;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
