--- src/main/java/com/nhnacademy/member_server/entity/Gender.java
@@ -1,7 +1,0 @@
-package com.nhnacademy.member_server.entity;
-
-public enum Gender {
-    MALE,
-    FEMALE,
-    UNKNOWN,
-}
