--- src/main/java/com/nhnacademy/member_server/entity/MemberPrincipal.java
@@ -1,46 +1,0 @@
-package com.nhnacademy.member_server.entity;
-
-import java.io.Serializable;
-import java.util.Collection;
-import java.util.List;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-@Getter
-@AllArgsConstructor
-public class MemberPrincipal implements UserDetails {
-    private Long memberId;
-    private String loginId;
-    private String role;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of(new SimpleGrantedAuthority(role));
-    }
-
-    @Override
-    public String getPassword() {
-        return null; // 패스워드 안 사용 시 null
-    }
-
-    @Override
-    public String getUsername() {
-        return loginId; // identifier로 사용됨
-    }
-
-    @Override
-    public boolean isAccountNonExpired() { return true; }
-
-    @Override
-    public boolean isAccountNonLocked() { return true; }
-
-    @Override
-    public boolean isCredentialsNonExpired() { return true; }
-
-    @Override
-    public boolean isEnabled() { return true; }
-}
