--- src/main/java/com/nhnacademy/member_server/entity/Role.java
@@ -1,6 +1,0 @@
-package com.nhnacademy.member_server.entity;
-
-public enum Role {
-    USER,
-    ADMIN
-}
