--- src/main/java/com/nhnacademy/member_server/entity/Status.java
@@ -1,7 +1,0 @@
-package com.nhnacademy.member_server.entity;
-
-public enum Status {
-    ACTIVE,
-    DORMANT,
-    WITHDRAWAL
-}
