+++ src/main/java/com/nhnacademy/member_server/entity/cartEntity/Cart.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.member_server.entity.cartEntity;
 
-import com.nhnacademy.member_server.entity.Member;
+import com.nhnacademy.member_server.entity.member.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
