+++ src/main/java/com/nhnacademy/member_server/service/impl/member/AuthServiceImpl.java
@@ -3,10 +3,7 @@ package com.nhnacademy.member_server.service.impl.member;
 import com.nhnacademy.member_server.dto.message.CouponIssueMessage;
 import com.nhnacademy.member_server.dto.request.member.MemberCreateRequest;
 import com.nhnacademy.member_server.dto.response.member.TokenDto;
-import com.nhnacademy.member_server.entity.Grade;
-import com.nhnacademy.member_server.entity.Member;
-import com.nhnacademy.member_server.entity.Role;
-import com.nhnacademy.member_server.entity.Status;
+import com.nhnacademy.member_server.entity.member.*;
 import com.nhnacademy.member_server.global.jwt.JwtUtil;
 import com.nhnacademy.member_server.repository.GradeRepository;
 import com.nhnacademy.member_server.repository.MemberRepository;
@@ -14,7 +11,9 @@ import com.nhnacademy.member_server.security.UserDetailsImpl;
 import com.nhnacademy.member_server.service.member.AuthService;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
+import java.util.UUID;
 import java.util.concurrent.TimeUnit;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
@@ -39,6 +38,7 @@ public class AuthServiceImpl implements AuthService {
     private final PasswordEncoder passwordEncoder;
     private final GradeRepository gradeRepository;
     private final RabbitTemplate rabbitTemplate;
+//    private final SocialLoginFactory socialLoginFactory;
 
     @Value("${jwt.refresh_expiration_time}")
     private Long refreshExpirationTime;
@@ -59,7 +59,7 @@ public class AuthServiceImpl implements AuthService {
         Long memberId = userDetails.getMember().getId();
         Role role = userDetails.getMember().getRole();
 
-        String accessToken = jwtUtil.createAccessToken(memberId, role, loginId);
+        String accessToken = jwtUtil.createAccessToken(memberId, role);
         String refreshToken = jwtUtil.createRefreshToken(memberId);
 
         redisTemplate.opsForValue().set(
@@ -144,7 +144,7 @@ public class AuthServiceImpl implements AuthService {
             throw new RuntimeException("탈퇴된 회원입니다.");
         }
 
-        String newAccessToken = jwtUtil.createAccessToken(member.getId(), member.getRole(), member.getLoginId());
+        String newAccessToken = jwtUtil.createAccessToken(member.getId(), member.getRole());
         String newRefreshToken = jwtUtil.createRefreshToken(member.getId());
 
         redisTemplate.opsForValue().set(
@@ -166,6 +166,66 @@ public class AuthServiceImpl implements AuthService {
         }
     }
 
-
+//    @Override
+//    public TokenDto loginSocial(String provider, String code) {
+//        SocialLoginStrategy strategy = socialLoginFactory.getStrategy(provider);
+//
+//        String providerId = strategy.getOAuth2MemberId(code);
+//
+//        Member member = memberRepository.findByProviderId(providerId).orElseGet(() -> {
+//            log.info("소셜 신규 회원. 자동 가입 {} / {}", provider, providerId);
+//            return socialSignup(provider, providerId);
+//        });
+//
+//        String accessToken = jwtUtil.createAccessToken(member.getId(), member.getRole());
+//        String refreshToken = jwtUtil.createRefreshToken(member.getId());
+//
+//        // 6. [Redis 저장] Refresh Token 저장
+//        redisTemplate.opsForValue().set(
+//                "RT:" + member.getId(),
+//                refreshToken,
+//                refreshExpirationTime,
+//                TimeUnit.MILLISECONDS
+//        );
+//
+//        // 7. 결과 반환
+//        return new TokenDto(accessToken, refreshToken);
+//
+//    }
+//
+//    private Member socialSignup(String provider, String providerId) {
+//        // 1. 비밀번호: 소셜 회원은 비번을 안 쓰지만, DB NotNull 제약 때문에 랜덤값 생성
+//        String randomPassword = passwordEncoder.encode(UUID.randomUUID().toString());
+//
+//        // 2. Login ID: 중복 안 되게 조합 (예: PAYCO_12345...)
+//        // (Tip: providerId가 길면 잘라쓰거나 그대로 써도 됨)
+//        String uniqueLoginId = provider + "_" + providerId;
+//
+//        // 3. 기본 등급 가져오기 (기존 코드 재사용)
+//        Grade basicGrade = gradeRepository.findByGradeName("GENERAL")
+//                .orElseGet(() -> gradeRepository.save(
+//                        Grade.builder().gradeName("GENERAL").min(0).pointRate(new BigDecimal("0.01")).build()
+//                ));
+//
+//        // 4. 회원 엔티티 생성 (빌더)
+//        Member member = Member.builder()
+//                .loginId(uniqueLoginId)
+//                .password(randomPassword)
+//                .name(provider + " User") // 닉네임 (추후 마이페이지에서 변경 유도)
+//                .email(uniqueLoginId + "@social.tmp") // 이메일 (임시값, 필요시 PAYCO Response에서 꺼내 써도 됨)
+//                .phone("010-0000-0000") // 전화번호 (필수라면 임시값)
+//                .birthDate(java.time.LocalDate.now()) // 생일 (임시값)
+//                .status(Status.ACTIVE)
+//                .role(Role.USER)
+//                .currentPoint(0L)
+//                .grade(basicGrade)
+//                .lastLoginAt(java.time.LocalDateTime.now())
+//                .provider(provider)      // "PAYCO"
+//                .providerId(providerId)  // 식별자
+//                .gender(Gender.UNKNOWN)  // (필요시 추가)
+//                .build();
+//
+//        return memberRepository.save(member);
+//    }
 
 }
