+++ src/main/java/com/nhnacademy/member_server/service/member/MemberService.java
@@ -3,7 +3,7 @@ package com.nhnacademy.member_server.service.member;
 import com.nhnacademy.member_server.dto.request.member.MemberUpdateRequest;
 import com.nhnacademy.member_server.dto.response.member.MemberResponse;
 import com.nhnacademy.member_server.dto.response.member.SimpleMemberResponse;
-import com.nhnacademy.member_server.entity.Role;
+import com.nhnacademy.member_server.entity.member.Role;
 
 import java.util.List;
