+++ src/test/java/com/nhnacademy/member_server/service/PointConcurrencyTest.java
@@ -3,10 +3,10 @@ package com.nhnacademy.member_server.service;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.nhnacademy.member_server.dto.request.PointTransactionRequest;
-import com.nhnacademy.member_server.entity.Grade;
-import com.nhnacademy.member_server.entity.Member;
-import com.nhnacademy.member_server.entity.Role;
-import com.nhnacademy.member_server.entity.Status;
+import com.nhnacademy.member_server.entity.member.Grade;
+import com.nhnacademy.member_server.entity.member.Member;
+import com.nhnacademy.member_server.entity.member.Role;
+import com.nhnacademy.member_server.entity.member.Status;
 import com.nhnacademy.member_server.repository.GradeRepository;
 import com.nhnacademy.member_server.repository.MemberRepository;
 import java.math.BigDecimal;
