+++ src/test/java/com/nhnacademy/member_server/service/PointServiceImplTest.java
@@ -13,13 +13,13 @@ import com.nhnacademy.member_server.dto.request.PointTransactionRequest;
 import com.nhnacademy.member_server.dto.response.PointAdminPolicyResponse;
 import com.nhnacademy.member_server.dto.response.PointBalanceResponse;
 import com.nhnacademy.member_server.dto.response.PointHistoryResponse;
-import com.nhnacademy.member_server.entity.Grade;
-import com.nhnacademy.member_server.entity.Member;
+import com.nhnacademy.member_server.entity.member.Grade;
+import com.nhnacademy.member_server.entity.member.Member;
 import com.nhnacademy.member_server.entity.PointEventType;
 import com.nhnacademy.member_server.entity.PointHistory;
 import com.nhnacademy.member_server.entity.PointPolicy;
-import com.nhnacademy.member_server.entity.Role;
-import com.nhnacademy.member_server.entity.Status;
+import com.nhnacademy.member_server.entity.member.Role;
+import com.nhnacademy.member_server.entity.member.Status;
 import com.nhnacademy.member_server.exception.BusinessException;
 import com.nhnacademy.member_server.exception.ErrorCode;
 import com.nhnacademy.member_server.repository.MemberRepository;
